+++ src/main/java/insane96mcp/survivalreimagined/mixin/DigDurabilityEnchantmentMixin.java
@@ -12,18 +12,14 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 @Mixin(DigDurabilityEnchantment.class)
 public abstract class DigDurabilityEnchantmentMixin {
 	@Inject(at = @At(value = "RETURN", ordinal = 1), method = "shouldIgnoreDurabilityDrop", cancellable = true)
-	private static void onGetMaxLevel(ItemStack stack, int lvl, RandomSource random, CallbackInfoReturnable<Boolean> cir) {
+	private static void onChanceToIgnoreDurabilityDrop(ItemStack stack, int lvl, RandomSource random, CallbackInfoReturnable<Boolean> cir) {
 		if (EnchantmentsFeature.isUnbreakingOverhaul()) {
 			//Shamelessly stolen from TinkersConstruct/src/main/java/slimeknights/tconstruct/tools/modifiers/upgrades/general/ReinforcedModifier.java
-			float chance;
+			float chance = 0.6f;
 			if (lvl < 5) {
-				// formula gives 25%, 45%, 60%, 70%, 75% for first 5 levels
-				chance = 0.025f * lvl * (11 - lvl);
-			}
-			else {
-				// after level 5.5 the above formula breaks, so just do +5% per level
-				// means for levels 6 to 10, you get 80%, 85%, 90%, 95%, 100%
-				chance = 0.75f + (lvl - 5) * 0.05f;
+				// formula gives 20%, 36%, 48%, 56%, 60% for first 5 levels
+				// In terms of durability, the tool lasts for x1.25, x1.56, x1.92, x2.27, x2.5 times more
+				chance = 0.02f * lvl * (11 - lvl);
 			}
 			cir.setReturnValue(random.nextFloat() < chance);
 		}