+++ src/main/java/insane96mcp/survivalreimagined/module/items/feature/CopperToolsExpansion.java
@@ -48,7 +48,7 @@ import java.util.List;
 public class CopperToolsExpansion extends Feature {
 	public static final TagKey<Item> COATED_EQUIPMENT = TagKey.create(Registries.ITEM, new ResourceLocation(SurvivalReimagined.MOD_ID, "equipment/coated_copper"));
 
-	public static final ILItemTier COPPER_ITEM_TIER = new ILItemTier(1, 143, 8f, 1.0f, 9, () -> Ingredient.of(Items.COPPER_INGOT));
+	public static final ILItemTier COPPER_ITEM_TIER = new ILItemTier(1, 65, 8f, 1.0f, 9, () -> Ingredient.of(Items.COPPER_INGOT));
 
 	public static final RegistryObject<Item> COPPER_SWORD = SRItems.REGISTRY.register("copper_sword", () -> new SwordItem(COPPER_ITEM_TIER, 3, -2.4F, new Item.Properties()));
 	public static final RegistryObject<Item> COPPER_SHOVEL = SRItems.REGISTRY.register("copper_shovel", () -> new ShovelItem(COPPER_ITEM_TIER, 1.5F, -3.0F, new Item.Properties()));